Streaming result
**Example 1**:
If I run
```
using(var session = driver.Session())
{
    var result = session.Run("RETURN 1");
}
```
The sequence of messages to exchange between client and server are:
```
-> INIT
<- SUCC_init

-> RUN
-> PULL_ALL
<- SUCC_run
```
After `}` the `session.Dispose()` will result in message sequence:
```
-> RESET
<- RECORD 1
<- SUCC_pull_all
<- SUCC_reset
```

So the RUN, PULL_ALL messages are eagerly sent, the SUCC for RUN is also eagerly received, but for records, it will only be retrieved in `result.Consume()` or similar iterations. If no result is iterated before the session is disposed, then the `RESET` message will force to sync all unhandled messages to make sure the session is clean when it returned to pool.

**Example 2**:
When I run
```
using(var session = driver.Session())
{
    session.Run("RETURN 1");
    var result = session.Run("RETURN 2");
    result.ToList();
}
```
Then the sequence of the messages are:
```
-> INIT
<- SUCC_init

 // run(return 1)
-> RUN
-> PULL_ALL
<- SUCC_run

// run(return 2)
-> RUN'
-> PULL_ALL'
<- RECORD 1
<- SUCC_pull_all
<- SUCC_run'

// result.ToList
<- RECORD 2
<-SUCC_pull_all'
```
In this case, as we did not save the result of the first `Run`, so they will just get lost if the second `Run` start to consume the result first. (We will improve this to cleverly buffer the results latter)

**Example 3**:
If we have
```
using(var session = driver.Session())
{
    try{session.Run("Invalid");}
    catch{}
    session.Run("RETURN 1");
}
```
Then the sequence of messages for the first `Run` are:
```
-> INIT
<- SUCC_init

-> RUN
-> PULL_ALL
<- FAIL_run
<- IGNORE_pull_all
-> ACK_FAIL
```
Note, as we failed before streaming, on the `RUN` message, we will then give up streaming but just drain all the messages that wait for the reply immediately. Therefore we will also wait for `IGNORE` message for `PULL_ALL` message. Finally the `ACK_FAIL` is added so that we will try to reset the session back to a good state for the following statement.

Then the messages for the second `Run` is the very same as the first example:
```
<- SUCC_ack_fail
-> RUN
-> PULL_ALL
<- SUCC_run

-> RESET
<- RECORD 1
<- SUCC_pull_all
<- SUCC_reset
```

In the new streaming design, there are two paths to receive messages from the server.
One is SocketClient.Receive (Session.Run()), another is iterator in RecordSet (result.Next()).
The one in RecordSet is pulled out for streaming records in result.

When handing failures, we also need to consider the two paths.
If an error happens before we start streaming records, then we will drain all the unhandled message queue and then throw the exception to the user to report the failure
INIT (sync), RUN, PULL_ALL (syncRun)
SUCC,        FAIL,IGNORE

However when the error arises after we start streaming the records, if the error is seen in streaming, then exception will be thrown immediately and the streaming will be stopped.
Any unhandled messages in the queue will not be addressed until ACK_FAIL or RESET is sent next time.
If the error is not seen during streaming, then the exception will be thrown when RESET is sent (in session.Dispose()).
INIT (sync), RUN, PULL_ALL (syncRun), (result.Consume()),RESET(sync)
SUCC,        SUCC,                     RECORD, FAIL,     SUCC

On Reset, all pending messages will be synced to ensure all records from the current session are consumed