Fixed the tests after new streaming
In the new streaming design, there are two paths to receive messages from the server. One is receive call in SocketClient (Session.Run()), another is record iterator in RecordSet (result.Next()). The one in RecordSet is added for streaming records in result.

When handing failures, we also need to consider the two paths.
If an error happens before we start streaming records, then we will drain all the unhandled message queue and then throw the exception to the user to report the failure
INIT (sync), RUN, PULL_ALL (syncRun)
SUCC,        FAIL,IGNORE

However when the error arises after we start streaming the records, if the error is seen in streaming, then exception will be thrown immediately and the streaming will be stopped. Any unhandled messages in the queue will not be addressed until RESET is sent next time. If the error is not seen during streaming, then the exception will be thrown when RESET is sent (in session.Dispose()).
INIT (sync), RUN, PULL_ALL (syncRun), (result.Consume()),RESET(sync)
SUCC,        SUCC,                     RECORD, FAIL,     SUCC