Added `ISession#Reset` to kill long running statements
The killing of a long running statement in a session is achieved by sending a `RESET` message to the server.
This `RESET` message could only be sent via another thread as the current session thread is blocked by reading a result from server.
However `ISession#Reset is the only method that could be used concurrently in `ISession`.

Changed to always buffer the result if the result is not consumed immediately within the session
Change to only flush but not receive on `run`. Leave `result.next` to recive.