Close tx automatically after session reset Enforce check if any error occured due to reset but has not been consumed before adding more messages or flushing messages
Improved reset thread-safe

Session reset touches message queues and tx status. Therefore in order to make this method call thread-safe with all others in session,
we need to lock the code that modifies tx status or the code that performs a certain action depending on the current tx status.

Session, transaction: lock the code around tx status change and the code around action performed on a specific tx status
SocketConnection: lock the code that change the status of interrupted status and the code related to message queuing