Resolve Dns to IPAddresses for socket connections
When establish a socket connection, instead of using `(Uri.Host, Uri.Port)`, we will use `(IPAddress, port)` instead to avoid issues on some platform where socket cannot be established with dns addresses due to one to many mapping.

If the uri given is already an ip address, then the ip address will be used directly. Otherwise, we will perform a dns lookup to resolve to ip addresses. The resolved ip addresses will be put in an array where IPv6 addresses will be placed first. Then socket client will try to establish connections with the server's ip addresses in order until connection established successfully.

This change does not implement connection pool level dns resolution. The connection pool still holds a pool for each uri (could either be a dns or ip address). While this PR does not need to be modified when connection pool level dns lookup is added as the socket client will not do a dns resolve if the uri is already an ip address.

Fix for #164