Added two stub tests to cancel statement execution with reset message Added IT tests to cancel statement execution with reset message TODO: what's the expected behaviour after reset for a session and/or tx?
This PR changed the place where to ack failure

The idea to mute ack failure:
We should disallow client to send `ACKFAIL` if the client already sent `RESET` as `RESET` will clean all failure state on server side. Then if we still send `ACKFAIL` after `RESET`, the server will error that there is no error to ack. Therefore we need to mute `ACKFAIL` from a `RESET` is sent from client till the reply of `RESET` is received.

The reason we need to change the place of ack failure:
After we changed to pull as much as messages from chunked input, we might receive and handle more than one message in `SocketConnection.ReceiveOne` method. While we ack failure for any server error only at the end of this method. Then at this moment, we already receives the reply from the server. It is too late to let the `AckFailure` method know whether it can ack or not. Therefore we need to move `AckFailure` as the moment the failure accours.

AckFail is a protocol. Therefore when we receive an error, we will try to ack fail. Then the connection will help us to test if it can ack or not and enqueue the messages if it can ack.

PooledConnection will no longer handle ack failure, instead, it only cares about if the connection has any unrecoverable error or not. If it has, then this connection will be closed. If it does not, then it will be reused.

Fixed a bug in DelegatedConn that error from chained tasks passed to next call as an aggregated error rather than its base error.