Improved the old connection pool statistics. And it is called Metrics, as we only aim to providing raw measurements.
A new dependency HdrHistogram is introduced to better support histogram used in our metrics.

The metrics are collected via a listener added in connection pool. While the listener interface is not directly exposed to end user as we do not want to have surprises from user's code. Instead we allow user to inject a metrics collector where we register metrics results to.