Allow Transaction to be used remotely
I won't pretend to know exactly why this works or if there are better
ways to solve the issues, so take what I state here with a grain of
salt.

The Neo4jClient driver supports TransactionScope and promotion of a
transaction to a distributed transaction. TransactionScope is important
when using other transactions as well, e.g. SQL transactions.

When starting a SQL transaction ofter a Neo4j transaction is started,
the transaction have to be promoted to a distributed transaction.
Neo4jClient uses AppDomain to handle this (at least I think that's the
reason), and AppDomain requires MarshalByRefObject to work properly.

Without MarshalByRefObject, things just doesn't work. When trying to
debug why nothing gets commited, the debugger breaks with a message that
the Neo4j Transaction isn't serializable. Adding SerializableAttribute
wasn't possible as this was needed all the way down, and
ConcurrentDictionary (or something similar) wasn't marked as
Serializable.

Adding MarshalByRefObject fixes Neo4jClient, allowing the Bolt
transactions to be commited.