Adding IAsyncSession#SessionConfig
to retrive configuration back after a session is created.
However this field is read-only. It also cannot be used as a configuration for another session.
In order to use the "same configuration" among different sessions, a user should use a function such as
```
void ConfigureSession(SessionConfigBuilder builder) =>
    builder.WithDefaultAccessMode(AccessMode.Read)
           .WithDatabase("neo4j");

driver.Session(ConfigureSession);
```