Expansion of BoltProtocolVersion class to support packing/unpacking of shorts and bytes.
Socket now requests a number X of most recent supported versions from the protocol.  This give flexibility should we expand past 4 in the future, and allows all versions to be supported and listed in the protocol factory.

Socket client now supports new BoltProtocolVersion, this means that nowhere below the socket uses anything but this type.  No more ints.