Added new message format for bolt protocol 4.1
Added HelloMessage serializer class for protocol 4.1
Added a new interface IConnectionProviderWithRoutingContext.  This exposes a Property that stores the RoutingSettings for any connections that will be provided.
Updated IConnectionPool and LoadBalancer to use the IConnectionProviderWithRoutingContext interface.  This is so the RoutingSettings can be fed to the connections.
Updated classes that were affected by the interface changes where needed.