Testkit backend was not rolling back open transactions on a RetryNegative
If a client error occurred and a RetryNegative message was received, the testkit backend was not rolling back any open transactions on the supplied session.

Added a list of transaction IDs to the NewSession object type that is populated on a Write or Read transaction creation. This is used to invoke the rollbacks when the RetryableNegative is processed.