Reworking of the Transaction management in the TestkitBackend
There was an issue with the handling of results and the differences between transaction functions, RunAsync and BeginTransaction. Added functionality to process and store results for transaction functions, and for the others a result is stored to be consumed later. This is done because results do not live outside the scope of a transaction function and the records will be lost if not handled correctly. 

Also had to add a TransactionWrapper type so that a delegate could be supplied at object construction to process a transactions results, depending on the above requirements. This enables TransactionRun to handle both types results in the correct ways without knowing the details of the transaction it is using.