Testkit backend (#490)
* Adding ResultConsume to protocol and its factory

* Fix for bookmark handling in SessionLastBookmarks protocl object

* Testkit backend was not rolling back open transactions on a RetryNegative

If a client error occurred and a RetryNegative message was received, the testkit backend was not rolling back any open transactions on the supplied session.

Added a list of transaction IDs to the NewSession object type that is populated on a Write or Read transaction creation. This is used to invoke the rollbacks when the RetryableNegative is processed.

* Reworking of the Transaction management in the TestkitBackend

There was an issue with the handling of results and the differences between transaction functions, RunAsync and BeginTransaction. Added functionality to process and store results for transaction functions, and for the others a result is stored to be consumed later. This is done because results do not live outside the scope of a transaction function and the records will be lost if not handled correctly. 

Also had to add a TransactionWrapper type so that a delegate could be supplied at object construction to process a transactions results, depending on the above requirements. This enables TransactionRun to handle both types results in the correct ways without knowing the details of the transaction it is using.

* Reverted change  to PopulateRecords method.

* Fix for threading issue with an async delegate called in a foreach loop.

A list of tasks is now created for a delegate that wraps the async functionality. The await is now on when all those tasks are complete.