Feature testkit backend protocol expansion (#496)
* Added VerifyConnectivity protocol object

Added class and processing code to support the VerifyConnectivity TestKit protocol message.

* Added CheckMultiDBSupport protocol object

Added class and processing logic to hand CheckMultiDBSupport TestKit protocol messages.

* Added ResolverResolutionCompleted protocol type

Added new protocol type
First pass at using new type in Driver resolver callback mechanism.

* Refactor of the ProtocolFactory

Removed type enumeration and mapping to strings. Now uses object types only, with and hashmap of protocol object types to verify against. Greatly simplified the system.

* Improved strictness on protocol violations

An exception is now thrown if the json parser finds a field that is not part of the object type it is trying to create. This is to help prevent divergence of the TestKit backend from the frontend.

* Added support for test blacklist and supporting protocol objects

Added StartTest protocol object request and its two potential responses, RunTest and SkipTest.

Added TestBlackList static class. This contains a list of tests that are not to be run and a reason for each one.

* Fix to Uri passing in NewDriver address resolver

Improved code to handle test uris that are non Uri library type conformant.

Added domainNameResolverRegistered and connectionTimeoutMs to the NewDriverType so that the stricter Json parsing would not throw an error.

* Bug fixing for StartTest protocol process

Added new exception type for unrecognized protocol messages to tidy up reporting of that error type.

Fixed the payload contents of the RunTest and SkipTest reponse messages.

Added StartTest to recognized protocol object types in the factory

Added txMetaData to session transaction functions and supporting config builder code.

Updated the test blacklist to include namespaces in test names.

* Added timeout member to session read and write transactions 

Updated protocol objects SessionWriteTransaction and SessionReadTransaction to have a timeout member that is used in the config builder.