Fixing stub scripts to use new format for GetRoutingTableProcedure calls
* Update to Require Stub server Facts and Theorys

As a new default behaviour tests that require the stub server will no longer be run on versions after and including 4.3. Testkit is taking over these tests. Eventually the stub integration tests will be removed from the project.

* Fixed a misuse of the Skip property in the tests.

* Further bug fixes to try and get the tests to behave.

* Ensuring AccessMode key is always written in metadata, even when write.

* Reverting previous change.

Test did not work out so removing mode attribute when it is a write operation.

* Removed mode from routing table scripts

* Updating remaining scripts that contain old getRoutingTable call format