Reverting AutoCommit  on SessionRun to use SendAsync (#528)
* First phase of implementing Authorization error handling

* Begin Transaction no longer requires a bookmark to sync.

SyncAsync is always called now. Previously the message was just enqueued and any errors would be discovered on result consumption.

* Fix implementation of GetFeatures testkit protocol message and further AuthorizationError work. (#524)

* First phase of implementing Authorization error handling

* Begin Transaction no longer requires a bookmark to sync.

SyncAsync is always called now. Previously the message was just enqueued and any errors would be discovered on result consumption.

* Update to LDAP Authorization work

Disabled all LDAP tests

Updated unit tests to take account of the use of SyncAsync rather than SendAsync in protocols when auotcommiting

Improved handling of exceptions to remove Task.FromException from async methods as it is only required for synchronous use.

* testkit-backend: Writing error over the wire when ArgumentException occurs (#525)

This issue was not allowing to run the tests to check if the driver will fail to be created when there are url querystring params for a bolt (direct) connection.

* First phase of implementing Authorization error handling

* Disabling all LDAP Testkit tests using Blacklist

* Removed duplicate entry from Testkit Blacklist

* Reverting AutoCommit Session Run to use SendAsync

Until LDAP is debugged reverting this and its relevant unit tests back to the original form.

Co-authored-by: Antonio Barcélos <antonio.barcelos@neo4j.com>