Rework of connection Reauthorization
Simplified and fully implements the specification.
A connection will detect the AuthorizationError then call the connection verify class that will mark the connection for Reauth along with any other connections it knows about (basically the pool the connection is part of) that are the same age or older than the triggering connection.