Removal of incorrect stress test.
PoolWithTxFuncWithFailingConnections was incorrectly implemented and had started to fail reliably. The test attempted to simulate failing connections under load by calling connection.DestroyAsync. This closes the connection, cleans it up and disposes of the relevant objects. Once this was done the test would attempt to consumeAsync.... this would then fail with a disposed object exception. Destroying a connection in this way is not the same as a failing connection. It is just testing the object disposal in a way that can never happen and is completely broken.