Fix for new handshake and unit test.
The handshake would select the highest protocol version supplied to it by the server, when instead it should have been selecting the highest version supplied by the server that is less than or equal to the highest protocol version the driver knows about (it's current version).
Example: Server sends a list of 5.0 -> 5.8 protocol versions. The driver is on protocol 5.6. It would select 5.8 instead of 5.6.

The unit test checking for this condition also had an error in it. On the second protocol entry in sent by the 'server' it supplied a version instead of a range.  This meant that under certain circumstances the test would pass, which just so happened to be the case with the current versions when the handshake was added.